added select user query and refactored
